Put blazesym-dev BPF functionality into separate file system module
Move the BPF functionality from an inline defined module into a proper
file based one, now that we have other definitions in the crate as well.
